backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
python-dotenv==1.0.1
crewai[tools]==1.12.2
faster-whisper==1.0.3
yt-dlp==2024.11.18
langchain==0.3.0
langchain-openai==0.2.0
langchain-groq==0.2.0
langchain-community==0.3.0
PyPDF2==3.0.1
python-docx==1.1.2
reportlab==4.2.2
fpdf2==2.8.1
matplotlib>=3.8.0
seaborn>=0.13.0
python-socketio>=4.6.0
requests>=2.31.0
